Fix carloads/rate_usd int-truncation bug; add HF sync script + session notes
carloads and rate_usd were typed int, but real USDA/Freightos data reports
fractional values (e.g. carloads=811.5) -- int(str) raised ValueError and
silently dropped the whole record. Found live during this pipeline's
first-ever real run (2 of 154,192 raw carloading records dropped). Changed
both to float across schemas.py/normalizer.py/storage.py's pyarrow schema.
rate_usd fixed proactively (same bug class, not yet live-tested since
Freightos is still key-blocked).

Also adds upload_huggingface.py (mirrors financial-data-pipeline's pattern,
concatenates partitioned parquet per table before upload) and
CLAUDE_SESSION_NOTES.md for this repo.

+++ src/freight_rail_pipeline/models/normalizer.py
@@ -54,7 +54,7 @@ class DataNormalizer:
                 railroad=str(raw.get("railroad", raw.get("carrier", "unknown"))),
                 commodity=str(raw.get("commodity", raw.get("commodity_desc", "unknown"))),
                 traffic_type=raw.get("type") or raw.get("traffic_type"),
-                carloads=int(carloads_raw),
+                carloads=float(carloads_raw),
                 units=raw.get("units", "carloads"),
                 origin_region=raw.get("origin", raw.get("origin_region")),
                 destination_region=raw.get("destination", raw.get("destination_region")),
@@ -124,7 +124,7 @@ class DataNormalizer:
                 ),
                 trade_lane=str(raw.get("tradeRoute", raw.get("trade_lane", "unknown"))),
                 container_type=str(raw.get("containerType", raw.get("container_type", "40GP"))),
-                rate_usd=int(rate_usd),
+                rate_usd=float(rate_usd),
                 region=raw.get("region"),
                 raw_record=raw,
             )
